WIP get trace and search with 128-bit trace IDs
This is work-in-progress support of the last step in 128-bit trace
support. This change aims to retrieve traces by all 128-bits instead of
just the lower 64. This is important when using backends that don't
support searches by partial trace ID.

This also checks that group-by logic considers the high bits.
Specifically, this tests that traces that share lower 64-bit are not
merged together in search results.

Certain aspects of this change are in-flux:

For example, should the overloaded signature look like this?
`getTrace(long traceIdHigh, long traceId)`

Should the setting for grouping on 128-bits be a request parameter (so
that it can be set in the UI), or a server parameter?

There's also incidental change needed, for example #1379